+++ src/main/kotlin/civictech/deliberate/repository/convert/DegreeWritingConverter.kt
@@ -5,7 +5,7 @@ import org.springframework.core.convert.converter.Converter
 import org.springframework.data.convert.WritingConverter
 
 @WritingConverter
-class DegreeWritingConverter : Converter<Degree, Float> {
-    override fun convert(source: Degree): Float = source.value
+class DegreeWritingConverter : Converter<Degree, Double> {
+    override fun convert(source: Degree): Double = source.value
 
 }
